check presentational structure
