+++ packages/tc-schema-validator/validation/graphql-defs.js
@@ -1,0 +1,10 @@
+const { parse } = require('graphql');
+const sdk = require('./sdk');
+
+try {
+	parse(sdk.getGraphqlDefs().join('\n'))).not.toThrow()
+} catch (error) {
+	console.error('GraphQL schema construction failed')
+	console.error(error)
+	throw error;
+}
