+++ packages/tc-schema-validator/validation/string-patterns.js
